ref_system/requirements.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==3.2.0
cffi==1.15.1
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
isort==5.11.5
itsdangerous==2.1.2
Jinja2==3.1.4
limits==3.10.1
Mako==1.2.4
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.0.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.8
pytz==2024.2
PyYAML==6.0.1
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sqladmin==0.14.1
SQLAlchemy==2.0.36
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.16.0
WTForms==3.0.1
zipp==3.15.0
